pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["claudio <claudiogfilho8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.*"
fastapi = {extras = ["standard"], version = "^0.115.2"}
ignr = "^2.2"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.0"
alembic = "^1.13.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev todolist/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=todolist -vv'
post_test = 'coverage html'